+++ lib/events/handlers/index.ts
@@ -5,3 +5,4 @@ export * from "./timeline";
 export * from "./opportunities";
 export * from "./compass";
 export * from "./academic";
+export * from "./opportunityGraph";