+++ lib/opportunity-graph/index.ts
@@ -1,0 +1,4 @@
+export * from "./types";
+export * from "./ontology/OpportunityOntology";
+export * from "./matching/OpportunityMatcher";
+export * from "./engine/OpportunityGraphEngine";